+++ packages/text-input-react/documentation/TextAreaExample.tsx
@@ -1,10 +1,18 @@
 import { LabelVariant } from "@fremtind/jkl-core";
 import React, { useState, ChangeEvent, FC } from "react";
 import { ExampleComponentProps, ExampleKnobsProps } from "../../../doc-utils";
-import { TextArea } from "../src";
+import { TextArea } from "../src/TextArea";
 
 export const textAreaExampleKnobs: ExampleKnobsProps = {
-    boolProps: ["Ekspanderende", "Starter åpen", "Med teller", "Skjul progress", "Med hjelpetekst", "Med feil"],
+    boolProps: [
+        "Ekspanderende",
+        "Starter åpen",
+        "Med teller",
+        "Skjul progress",
+        "Med hjelpetekst",
+        "Med feil",
+        "Med tooltip",
+    ],
     choiceProps: [
         {
             name: "Variant",
@@ -27,6 +35,12 @@ export const TextAreaExample: FC<ExampleComponentProps> = ({ choiceValues, boolV
         ? "Du må fylle ut en beskrivelse. Beskriv så utfyllende som mulig."
         : undefined;
 
+    const tooltipProps = boolValues?.["Med tooltip"]
+        ? {
+              content: "Beskriv så utfyllende som mulig.",
+          }
+        : undefined;
+
     return (
         <TextArea
             className="jkl-spacing-xl--top"
@@ -34,10 +48,11 @@ export const TextAreaExample: FC<ExampleComponentProps> = ({ choiceValues, boolV
             name="beskrivelse"
             helpLabel={helpLabel}
             errorLabel={errorLabel}
+            tooltipProps={tooltipProps}
             autoExpand={autoExpand}
             startOpen={startOpen}
             counter={medTeller ? { maxLength: 200, hideProgress: skjulProgress } : undefined}
-            variant={variant}
+            labelProps={{ variant }}
             value={value}
             onChange={handleChange}
         />
@@ -51,7 +66,12 @@ export const textAreaExampleCode = ({ choiceValues, boolValues }: ExampleCompone
     helpLabel=${boolValues?.["Med hjelpetekst"] ? `"Beskriv så utfyllende som mulig"` : `{undefined}`}
     errorLabel=${
         boolValues?.["Med feil"] ? `"Du må fylle ut en beskrivelse. Beskriv så utfyllende som mulig."` : `{undefined}`
-    }
+    }${
+    boolValues?.["Med tooltip"]
+        ? `
+    tooltipProps={{ content: "Beskriv så utfyllende som mulig." }}`
+        : ""
+}
     autoExpand={${boolValues?.["Ekspanderende"]}}
     startOpen={${boolValues?.["Starter åpen"]}}
     counter={${