+++ packages/checkbox-react/CHANGELOG.md
@@ -3,6 +3,16 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# 9.0.0 (2023-01-16)
+
+### Features
+
+- forenkle bruken av tooltip med FieldGroup ([779eb91](https://github.com/fremtind/jokul/commit/779eb913fa992dac394ee0adf147c34b5b2ee78b))
+
+### BREAKING CHANGES
+
+- `variant` er fjernet. Bruk `labelProps={{ variant }}` i stedet.
+
 ## 8.0.11 (2022-11-22)
 
 ### Bug Fixes